Optimize day 5
Did the math on ranges as I mentioned in the previous commit.
So whenever a range crosses a translation's boundary, we create a
new range. We're not combining them, so on an infinitely large
problem we would get progressively slower. However for this input it's
been reduced from 83 seconds to 0.5 milliseconds.

We're still checking all transition ranges for every number, since sorting
each and looping through them only when it was within bounds turned out
to be slower.